% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Get balance of minimisation}
\usage{
balance(object)
}
\arguments{
\item{object}{An object of class `mini`}
}
\value{
An object of class `balance.mini`
}
\description{
Get balance minimisation randomisation
}
