% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimise.R
\name{minimise}
\alias{minimise}
\title{Minimisation for a dataset}
\usage{
minimise(
  data,
  groups = 3,
  factors,
  burnin = 10,
  minprob = 0.8,
  ratio = rep(1, groups),
  group.names = NULL,
  seed = NULL,
  check.eligibility = FALSE
)
}
\arguments{
\item{data}{a \code{data.frame} object with one line per participant and columns
for the minimisation factors.}

\item{groups}{an integer, the number of groups to randomise to, default is 3}

\item{factors}{a character vector with the factors for minimisation}

\item{burnin}{an integer, the burnin length before minimisation kicks in,
default is 10. Must be > 0 and < total sample size.}

\item{minprob}{the minimisation probability. The default is to give 0.8
probability to the group which would lead to the least
imbalance.}

\item{ratio}{a numeric vector of randomisation ratios in ascending order
(must be of length equal to the number of groups)}

\item{group.names}{optional, a character vector with the group names, must be
the same length as \code{groups}.}

\item{seed}{optional, an integer that is used with \code{set.seed()} for
offsetting the random number generator.}

\item{check.eligibility}{logical indicating whether participants have varying
eligibility for different groups. If \code{TRUE} then the
column \code{eligible} should exist in the data.}
}
\value{
(Invisibly) the data.frame with an additional column \code{Group}
indicating numerically which group has been allocated.
}
\description{
\code{minimise} randomises patients in a data frame using minimisation
}
\examples{
# Minimisation to 3 groups, with two factors and a burnin of 15, using the
# patients data from the package
(mini <- minimise(patients, groups = 3, factors = c("sex", "stage"),
                  burnin = 15))

# View data with group info
as.data.frame(mini)

# Use 1:1:2 ratio
minimise(patients, groups = 3, factors = c("sex", "stage"), burnin = 5,
         ratio = c(1,1,2))

}
