% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimise_app.R
\name{minimise_app}
\alias{minimise_app}
\title{App interface for the multimini package}
\usage{
minimise_app(data, ...)
}
\arguments{
\item{data}{patient data}

\item{...}{options passed to \code{shiny::shinyApp()}}
}
\description{
App interface for the multimini package
}
