% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patients}
\alias{patients}
\title{Randomly Generated Patient Data}
\format{
\subsection{\code{patients}}{

A data frame with 150 rows and 4 columns:
\describe{
\item{sex}{Patient sex (M or F)}
\item{stage}{Stage of disease (I, II, or III)}
\item{site}{Recruitment site (1 to 10)}
\item{eligible}{Which treatments are they eligible for? (A, B, C)}
}
}
}
\usage{
patients
}
\description{
A randomly generated data set of "patient" data
}
\keyword{datasets}
