% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimise.R
\name{plot.mini}
\alias{plot.mini}
\title{Plots to explore minimisation results}
\usage{
\method{plot}{mini}(x, show.plots = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{mini} as a result of \code{minimises}}

\item{show.plots}{logical; if \code{FALSE} plots won't be displayed, useful when
assigning the plots for future use.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Plots to explore minimisation results
}
