% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mini.R
\name{simulate_mini}
\alias{simulate_mini}
\title{Simulation of minimisation}
\usage{
simulate_mini(
  data,
  Nsims = 100,
  groups = 3,
  factors,
  burnin,
  minprob,
  ratio = rep(1, groups)
)
}
\arguments{
\item{data}{a \code{data.frame} object with one line per participant and columns
for the minimisation factors.}

\item{Nsims}{an integer, the number of simulations to run per scenario,
default is 100.}

\item{groups}{an integer, the number of groups to randomise, default is 3.}

\item{factors}{a list of character vectors with the names of the factors for
minimisation}

\item{burnin}{a vector of integers of possible burnin lengths before
minimisation kicks in. Individual values must be > 0 and
< total sample size.}

\item{minprob}{a list of vectors (each of the same length as \code{groups}) with
the possible minimisation probabilities.}

\item{ratio}{a numeric vector of randomisation ratios (must be of length
equal to the number of groups).}
}
\description{
\code{simulate_mini} repeatedly minimises a data set of patients with different
specified burnin periods and/or minimisation probabilities. Its purpose is to
guide decisions when designing a randomisation strategy for a randomised
controlled trial.
}
