% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Get balance of minimisation}
\usage{
balance(object)
}
\arguments{
\item{object}{An object of class \code{mini}}
}
\value{
An object of class \code{balance.mini}
}
\description{
Get balance minimisation randomisation
}
\examples{
# Minimisation to 3 groups, with two factors and a burnin of 15
mini <- minimise(patients, groups = 3, factors = c("sex", "stage"),
                 burnin = 15)

balance(mini)

}
