% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimise.R
\name{minimise}
\alias{minimise}
\title{Minimisation for a dataset}
\usage{
minimise(
  data,
  groups = 3,
  factors,
  burnin = 10,
  minprob = c(0.8, rep(0.2/(groups - 1), groups - 1)),
  stratify = NULL
)
}
\arguments{
\item{data}{a `data.frame` object with one line per participant and columns
for the minimisation factors.}

\item{groups}{an integer, the number of groups to randomise to, default is 3}

\item{factors}{a character vector with the factors for minimisation}

\item{burnin}{an integer, the burnin length before minimisation kicks in,
default is 10}

\item{minprob}{a vector of the same length as `groups` with the minimisation
probabilities. The default is to give 0.8 probability to the
group which would lead to the least imbalance and to allocate
the remaining probability equally to the other groups.}

\item{stratify}{if stratification is to be used then a character string
specifying the name of the stratification variable (e.g.,
"site"). Default is `NULL` for no stratification.}
}
\value{
(Invisibly) the data.frame with an additional column `Group` indicating
numerically which group has been allocated.
}
\description{
`minimise` randomises patients in a data frame using minimisation
}
