% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mini.R
\name{simulate_mini}
\alias{simulate_mini}
\title{Simulation of minimisation}
\usage{
simulate_mini(
  sampsize,
  factors,
  Nsims = 100,
  groups = 3,
  burnin,
  minprob,
  ratio = rep(1, groups)
)
}
\arguments{
\item{sampsize}{the desired sample size of the data, i.e., the sample size
of your prospective trial.}

\item{factors}{a list of factors, each a list containing two items. The first
is \code{levels} which is either a vector of the level names OR the number of
levels. The second item is either \code{props}, a vector of proportions equal to
the number of levels in the factor; OR prop.dist which is a named vector
containing the mean and sd of the proportions.}

\item{Nsims}{an integer, the number of simulations to run per scenario,
default is 100.}

\item{groups}{an integer, the number of groups to randomise, default is 3.}

\item{burnin}{a vector of integers of possible burnin lengths before
minimisation kicks in. Individual values must be > 0 and
< total sample size.}

\item{minprob}{a vector of the possible minimisation probabilities.}

\item{ratio}{a numeric vector of randomisation ratios (must be of length
equal to the number of groups).}
}
\description{
\code{simulate_mini} repeatedly generates random data and then minimises with
different specified burnin periods and/or minimisation probabilities. Its
purpose is to guide decisions when designing a randomisation strategy for a
randomised controlled trial.
}
\examples{
simulate_mini(sampsize = 150,
              factors = list(sex = list(levels = c("M", "F"),
                                        props = c(0.5, 0.5)),
                             site = list(levels = 12,
                                         prop.dist = c(mean = 0.1, sd = 0.05))),
              Nsims = 10, groups = 3, burnin = c(10, 15),
              minprob = c(0.7, 0.8, 0.9), ratio = c(1,1,1))

}
