% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.mini}
\alias{update.mini}
\title{Update a minimisation with new data}
\usage{
\method{update}{mini}(object, new.data, ...)
}
\arguments{
\item{object}{an object of class mini, i.e., resulting from \code{minimise()}}

\item{new.data}{a \code{data.frame} object with one line per new participant and
columns for the minimisation factors. Must be of the same
format as the \code{data.frame} in \code{object} (minus \code{Group})}
}
\value{
(Invisibly) the data.frame of all participants with group existing
and new group allocations
}
\description{
\code{update} updates the minimisation with new participant data, using the
existing group allocations and factor information to allocate the new
participants using minimisation.
}
\examples{
mini <- minimise(patients, groups = 3, factors = c("sex", "stage"),
                 burnin = 10)

new.patients <- data.frame(
  sex = c("F", "F", "M"),
  stage = c("II","I", "I"),
  site = c(9, 9, 1)
)

update(mini, new.patients)

}
